BVAL-616 Extracting @PositiveOrZero/@NegativeOrZero from @Positive/@Negative
